Generate code to register custom resource mappers
Resource annotations (both local and global) allow to disable generating of mappers for the resource class and to instead reference custom implementation classes, instances or names which shall be used for injection in the init function.

This functionality is now implemented, and I simplified and cleaned up the codebase while implementing this.