moved reader/builder call construction from mustache template to dart code
The reader/builder calls already are quite involved and make the templates hard to read and maintain, but in future we
will support RdfMapEntry annotations and then it will be even worse. So we rather have that logic in dart code.

Note that we are not using 'IfNotNull'
builder method suffix any more, because 'when' is much more generic and allows to achieve the same - vor all possible methods, even if 'IfNotNull'  is not available.